BRDF

+++ Source/BRDF/BxDF.cpp
@@ -7,7 +7,7 @@ BxDF::~BxDF()
 
 }
 
-Spectrum BxDF::Sample_f( const Vector3f& wo , Vector3f* wi , const Point2f& samplePoint , float* pdf ) const
+Spectrum BxDF::Sample_f( const Vector3f& wo , Vector3f* wi , const Point2f& samplePoint , double* pdf ) const
 {
 	// 生成半球采样方向
 	*wi = CosineSampleHemisphere( samplePoint );
@@ -29,7 +29,7 @@ Spectrum BxDF::Sample_f( const Vector3f& wo , Vector3f* wi , const Point2f& samp
 // wi是位于物体的局部空间
 // 因此法线为(0 , 0 , 1)
 // 因此cos(theta) = (0 , 0 , 1) * wi = wi.z
-float BxDF::PDF( const Vector3f& wi , const Vector3f& wo ) const
+double BxDF::PDF( const Vector3f& wi , const Vector3f& wo ) const
 {
 	return ( wi.z * wo.z > 0.0 ) ? fabs( wi.z ) * INV_PI : 0.0f;
 }
@@ -46,8 +46,8 @@ Spectrum BxDF::rho( int nSamples , Point2f* Samples1 , Point2f* Samples2 ) const
 		// 生成半球内的随机采样方向
 		wo = UniformSampleHemisphere( Samples2[i] );
 
-		float PdfOutput = INV_TWO_PI; 
-		float PdfInput = 0.0f;
+		double PdfOutput = INV_TWO_PI; 
+		double PdfInput = 0.0f;
 
 		Spectrum f = Sample_f( wo , &wi , Samples1[i] , &PdfInput );
 
@@ -69,7 +69,7 @@ Spectrum BxDF::rho( const Vector3f& wo , int nSamples , Point2f* samples ) const
 	for( int i = 0; i < nSamples; i++ )
 	{
 		Vector3f wi;
-		float pdf = 0.0f;
+		double pdf = 0.0f;
 
 		Spectrum f = Sample_f( wo , &wi , samples[i] , &pdf );
 
@@ -79,5 +79,5 @@ Spectrum BxDF::rho( const Vector3f& wo , int nSamples , Point2f* samples ) const
 		}
 	}
 
-	return r / float( nSamples );
+	return r / double( nSamples );
 }
