+++ Source/Filter/BoxFilter.cpp
@@ -1,7 +1,7 @@
 #include "PCH.h"
 #include "BoxFilter.h"
 
-float BoxFilter::Evalute( float x , float y ) const
+double BoxFilter::Evalute( double x , double y ) const
 {
 	return 1;
 }
