+++ Source/Math/Matrix4.h
@@ -12,14 +12,14 @@ class Matrix4f
 public:
 	inline Matrix4f();
 
-	inline Matrix4f( const float arr[4][4] );
+	inline Matrix4f( const double arr[4][4] );
 
-	inline Matrix4f( const float arr[16] );
+	inline Matrix4f( const double arr[16] );
 
-	inline Matrix4f( float _m00 , float _m01 , float _m02 , float _m03 ,
-					 float _m10 , float _m11 , float _m12 , float _m13 ,
-					 float _m20 , float _m21 , float _m22 , float _m23 ,
-					 float _m30 , float _m31 , float _m32 , float _m33 );
+	inline Matrix4f( double _m00 , double _m01 , double _m02 , double _m03 ,
+					 double _m10 , double _m11 , double _m12 , double _m13 ,
+					 double _m20 , double _m21 , double _m22 , double _m23 ,
+					 double _m30 , double _m31 , double _m32 , double _m33 );
 
 	inline Matrix4f( const Matrix4f& matrix );
 
@@ -44,19 +44,19 @@ public:
 	static Matrix4f Mul( const Matrix4f& lhs , const Matrix4f& rhs );
 
 public:
-	static Matrix4f Rotate( float fThetaX , float fThetaY , float fThetaZ );
+	static Matrix4f Rotate( double fThetaX , double fThetaY , double fThetaZ );
 
-	static Matrix4f RotateX( float fTheta );
+	static Matrix4f RotateX( double fTheta );
 
-	static Matrix4f RotateY( float fTheta );
+	static Matrix4f RotateY( double fTheta );
 
-	static Matrix4f RotateZ( float fTheta );
+	static Matrix4f RotateZ( double fTheta );
 
-	static Matrix4f RotateAxis( const Vector3f& axis , float fTheta );
+	static Matrix4f RotateAxis( const Vector3f& axis , double fTheta );
 
-	static Matrix4f Scale( float fScaleX = 1.0f , float fScaleY = 1.0f , float fScaleZ = 1.0f );
+	static Matrix4f Scale( double fScaleX = 1.0f , double fScaleY = 1.0f , double fScaleZ = 1.0f );
 
-	static Matrix4f Translate( float fx = 0.0f , float fy = 0.0f , float fz = 0.0f );
+	static Matrix4f Translate( double fx = 0.0f , double fy = 0.0f , double fz = 0.0f );
 
 public:
 	friend Point3f operator* ( const Point3f& p , const Matrix4f& matrix );
@@ -67,13 +67,13 @@ public:
 	{
 		struct
 		{
-			float m00 , m01 , m02 , m03;
-			float m10 , m11 , m12 , m13;
-			float m20 , m21 , m22 , m23;
-			float m30 , m31 , m32 , m33;
+			double m00 , m01 , m02 , m03;
+			double m10 , m11 , m12 , m13;
+			double m20 , m21 , m22 , m23;
+			double m30 , m31 , m32 , m33;
 		};
 
-		float m[16];
+		double m[16];
 	};
 };
 #endif
